[refactor] pen header independent
